automatically detect if node is present (and use it if it is), refs #12673
git-svn-id: http://svn.dojotoolkit.org/src/util/trunk@24967 560b804f-0ae3-0310-86f3-f6aa0a117693

+++ buildscripts/build.sh
@@ -1,7 +1,8 @@
 #!/bin/sh
 
-# for rhino...
-java -Xms256m -Xmx256m  -cp ../shrinksafe/js.jar:../closureCompiler/compiler.jar:../shrinksafe/shrinksafe.jar org.mozilla.javascript.tools.shell.Main  ../../dojo/dojo.js baseUrl=../../dojo load=build "$@"
-
-# for node...
-#node ../../dojo/dojo.js load=build "$@"
+if which -s node
+then
+   node ../../dojo/dojo.js load=build "$@"
+else
+   java -Xms256m -Xmx256m  -cp ../shrinksafe/js.jar:../closureCompiler/compiler.jar:../shrinksafe/shrinksafe.jar org.mozilla.javascript.tools.shell.Main  ../../dojo/dojo.js baseUrl=../../dojo load=build "$@"
+fi
